requirements.txt
google-genai
pydantic
python-dotenv
fastapi
uvicorn
googlesearch-python
beautifulsoup4
requests
